input works for both stdin and argument cases, League class moved to it's own file
