+++ League.py
@@ -1,0 +1,11 @@
+class League:
+    def __init__(self):
+        self.teams = {}
+        self.table = []
+
+    def print_table(self):
+        return
+
+    def fill_league(self, matches):
+        for match in matches:
+            print (match)
